requirements.txt
fastapi[standard]
langgraph
langchain-openai
python-dotenv
openai
requests
feedparser
beautifulsoup4
lxml